# [ ROCm ] tuned MultiTensorApply.cuh parameters for better perf on _foreach_addcdiv_