This configuration helps on ROCm (MI355) in the operator _foreach_addcdiv_.

before:
torch.float32 time (avg): 25.87 ms
torch.bfloat16 time (avg): 14.1 ms

after:
torch.float32 time (avg): 20.4 ms
torch.bfloat16 time (avg): 10.3 ms

Repro file: 
[foreach_addcdiv_repro.py](https://github.com/user-attachments/files/22549792/foreach_addcdiv_repro.py)